requirements.txt
# torch==2.4.0
# torchvision==0.19.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.13.2
transformers==4.43.3
bitsandbytes==0.43.2
sentencepiece==0.2.0
accelerate==0.33.0
protobuf==5.27.2
open_clip_torch==2.26.1
openai-clip==1.0.1
nltk==3.8.1
scipy==1.10.1

